Modify ICLPayload
• Move mutable class to same header file as immutable class
• Change type of contentLength to unsigned long long
• Modify NSCoder calls to be more type sensitive

+++ Classes (Shared)/Headers/Internal/ICLPayloadInternal.h
@@ -40,7 +40,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface ICLPayload ()
 {
 @protected
-	NSUInteger _contentLength;
+	unsigned long long _contentLength;
 	NSSize _contentSize;
 	NSArray<NSURL *> *_styleResources;
 	NSArray<NSURL *> *_scriptResources;
